trained crappy classifier

+++ link_bot_data/src/link_bot_data/classifier_dataset.py
@@ -60,7 +60,7 @@ class ClassifierDataset(BaseDataset):
     def post_process(self, dataset: tf.data.TFRecordDataset, n_parallel_calls: int):
         def _add_time(example: Dict):
             # this function is called before batching occurs, so the first dimension should be time
-            example['time'] = example[add_predicted(self.state_keys[0])].shape[0]
+            example['time'] = tf.cast(example[add_predicted(self.state_keys[0])].shape[0], tf.int64)
             return example
 
         dataset = dataset.map(_add_time)
